+++ apps/bfd-server/bfd-server-war/src/main/java/gov/cms/bfd/server/war/FDADrugDataUtilityApp.java
@@ -1,27 +1,11 @@
 package gov.cms.bfd.server.war;
 
-import java.io.BufferedOutputStream;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.net.URL;
-import java.nio.charset.Charset;
-import java.nio.charset.CharsetDecoder;
-import java.nio.charset.CharsetEncoder;
-import java.nio.charset.CodingErrorAction;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Comparator;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-import org.apache.commons.io.FileUtils;
 
 /**
  * A simple application that downloads the FDA NDC (national drug code) file; unzips it and then
@@ -78,7 +62,7 @@ public final class FDADrugDataUtilityApp {
     Path convertedNdcDataFile = outputPath.resolve(FDA_PRODUCTS_RESOURCE);
     if (!Files.exists(convertedNdcDataFile)) {
       try {
-        buildProductsResource(convertedNdcDataFile, workingDir);
+        DataUtilityCommons.buildProductsResource(convertedNdcDataFile, workingDir);
       } finally {
         // Recursively delete the working dir.
         Files.walk(workingDir)
@@ -88,105 +72,5 @@ public final class FDADrugDataUtilityApp {
             .forEach(File::delete);
       }
     }
-  }
-
-  /**
-   * Creates the {@link #FDA_PRODUCTS_RESOURCE} file in the specified location.
-   *
-   * @param convertedNdcDataFile the output file/resource to produce
-   * @param workingDir a directory that temporary/working files can be written to
-   * @throws IOException (any errors encountered will be bubbled up)
-   */
-  private static void buildProductsResource(Path convertedNdcDataFile, Path workingDir)
-      throws IOException {
-    // download FDA NDC file
-    Path downloadedNdcZipFile =
-        Paths.get(workingDir.resolve("ndctext.zip").toFile().getAbsolutePath());
-    URL ndctextZipUrl = new URL("https://www.accessdata.fda.gov/cder/ndctext.zip");
-    if (!Files.isReadable(downloadedNdcZipFile)) {
-      // connectionTimeout, readTimeout = 10 seconds
-      FileUtils.copyURLToFile(
-          ndctextZipUrl, new File(downloadedNdcZipFile.toFile().getAbsolutePath()), 10000, 10000);
-    }
-
-    // unzip FDA NDC file
-    unzip(downloadedNdcZipFile, workingDir);
-    Path originalNdcDataFile = workingDir.resolve("product.txt");
-    if (!Files.isReadable(originalNdcDataFile))
-      originalNdcDataFile = workingDir.resolve("Product.txt");
-    if (!Files.isReadable(originalNdcDataFile))
-      throw new IllegalStateException("Unable to locate product.txt in " + ndctextZipUrl);
-
-    // convert file format from cp1252 to utf8
-    CharsetDecoder inDec =
-        Charset.forName("windows-1252")
-            .newDecoder()
-            .onMalformedInput(CodingErrorAction.REPORT)
-            .onUnmappableCharacter(CodingErrorAction.REPORT);
-
-    CharsetEncoder outEnc =
-        StandardCharsets.UTF_8
-            .newEncoder()
-            .onMalformedInput(CodingErrorAction.REPORT)
-            .onUnmappableCharacter(CodingErrorAction.REPORT);
-
-    try (FileInputStream is = new FileInputStream(originalNdcDataFile.toFile().getAbsolutePath());
-        BufferedReader reader = new BufferedReader(new InputStreamReader(is, inDec));
-        FileOutputStream fw =
-            new FileOutputStream(convertedNdcDataFile.toFile().getAbsolutePath());
-        BufferedWriter out = new BufferedWriter(new OutputStreamWriter(fw, outEnc))) {
-
-      for (String in; (in = reader.readLine()) != null; ) {
-        out.write(in);
-        out.newLine();
-      }
-    }
-  }
-
-  /**
-   * Extracts a zip file specified by the zipFilePath to a directory specified by destDirectory
-   * (will be created if does not exists)
-   *
-   * @param zipFilePath
-   * @param destDirectory
-   * @throws IOException
-   */
-  private static void unzip(Path zipFilePath, Path destDirectory) throws IOException {
-    ZipInputStream zipIn =
-        new ZipInputStream(new FileInputStream(zipFilePath.toFile().getAbsolutePath()));
-    ZipEntry entry = zipIn.getNextEntry();
-    // iterates over entries in the zip file
-    while (entry != null) {
-      Path filePath = Paths.get(destDirectory.toFile().getAbsolutePath(), entry.getName());
-      if (!entry.isDirectory()) {
-        // if the entry is a file, extracts it
-        extractFile(zipIn, filePath);
-      } else {
-        // if the entry is a directory, make the directory
-        File dir = new File(filePath.toFile().getAbsolutePath());
-        dir.mkdir();
-      }
-      zipIn.closeEntry();
-      entry = zipIn.getNextEntry();
-    }
-    zipIn.close();
-  }
-
-  /**
-   * Extracts a zip entry (file entry)
-   *
-   * @param zipIn
-   * @param filePath
-   * @throws IOException
-   */
-  private static void extractFile(ZipInputStream zipIn, Path filePath) throws IOException {
-    BufferedOutputStream bos =
-        new BufferedOutputStream(new FileOutputStream(filePath.toFile().getAbsolutePath()));
-    byte[] bytesIn = new byte[BUFFER_SIZE];
-    int read = 0;
-    while ((read = zipIn.read(bytesIn)) != -1) {
-      bos.write(bytesIn, 0, read);
-    }
-    bos.close();
   }
 }
